+++ test/Sentry.Unity.Tests/UnityEventScopeTests.cs
@@ -163,7 +163,7 @@ public sealed class UnityEventProcessorThreadingTests
         Assert.AreEqual(systemInfo.TargetFrameRate!.Value, unityContext!.TargetFrameRate);
         Assert.AreEqual(systemInfo.CopyTextureSupport!.Value, unityContext.CopyTextureSupport);
         Assert.AreEqual(systemInfo.RenderingThreadingMode!.Value, unityContext.RenderingThreadingMode);
-        Assert.AreEqual(SceneManager.GetActiveScene().name, unityContext.ActiveSceneName);
+        Assert.AreEqual(captureOnUiThread ? SceneManager.GetActiveScene().name : null, unityContext.ActiveSceneName);
 
         Assert.IsNull(@event.ServerName);
     }
@@ -338,7 +338,8 @@ public sealed class UnityEventProcessorTests
 
         var sentryOptions = new SentryUnityOptions { SendDefaultPii = true };
         var scopeUpdater = new UnityScopeUpdater(sentryOptions, _testApplication);
-        var unityEventProcessor = new UnityEventProcessor(sentryOptions);
+        var unityInfo = new TestUnityInfo { IL2CPP = true };
+        var unityEventProcessor = new UnityEventProcessor(sentryOptions, unityInfo);
         var scope = new Scope(sentryOptions);
         var sentryEvent = new SentryEvent();
         var transaction = new SentryTransaction("name", "operation");
@@ -438,7 +439,6 @@ public sealed class UnityEventProcessorTests
     [TestCase(false)]
     public void UnityProtocol_Assigned(bool isIL2CPP)
     {
-        var sceneManager = new SceneManagerIntegrationTests.FakeSceneManager { ActiveSceneName = "TestScene" };
         var systemInfo = new TestSentrySystemInfo
         {
             EditorVersion = "TestEditorVersion2022.3.2f1",
@@ -450,7 +450,7 @@ public sealed class UnityEventProcessorTests
         MainThreadData.SentrySystemInfo = systemInfo;
         MainThreadData.CollectData();
 
-        var sut = new UnityScopeUpdater(_sentryOptions, _testApplication, new TestUnityInfo { IL2CPP = isIL2CPP }, sceneManager);
+        var sut = new UnityScopeUpdater(_sentryOptions, _testApplication);
         var scope = new Scope(_sentryOptions);
 
         // act
@@ -465,7 +465,6 @@ public sealed class UnityEventProcessorTests
         Assert.AreEqual(systemInfo.TargetFrameRate!.Value, unityProtocol.TargetFrameRate);
         Assert.AreEqual(systemInfo.CopyTextureSupport!.Value, unityProtocol.CopyTextureSupport);
         Assert.AreEqual(systemInfo.RenderingThreadingMode!.Value, unityProtocol.RenderingThreadingMode);
-        Assert.AreEqual(isIL2CPP ? sceneManager.GetActiveScene().Name : null, unityProtocol.ActiveSceneName);
     }
 
     [Test]
@@ -564,6 +563,28 @@ public sealed class UnityEventProcessorTests
 
         // assert
         Assert.IsNull(scope.Contexts.Gpu.GraphicsShaderLevel);
+    }
+
+    [Test]
+    [TestCase(true, true)]
+    [TestCase(false, true)]
+    [TestCase(true, false)]
+    [TestCase(false, false)]
+    public void Process_SetsActiveSceneName(bool isEditor, bool isIL2CPP)
+    {
+        var sentryOptions = new SentryUnityOptions();
+        var application = new TestApplication { IsEditor = isEditor };
+        var unityInfo = new TestUnityInfo { IL2CPP = isIL2CPP };
+        var sceneManager = new SceneManagerIntegrationTests.FakeSceneManager { ActiveSceneName = "TestScene" };
+        var sut = new UnityEventProcessor(sentryOptions, unityInfo, application, sceneManager);
+        var sentryEvent = new SentryEvent();
+
+        sut.Process(sentryEvent);
+
+        sentryEvent.Contexts.TryGetValue(Unity.Protocol.Unity.Type, out var unityProtocolObject);
+        var unityProtocol = unityProtocolObject as Unity.Protocol.Unity;
+        Assert.NotNull(unityProtocol);
+        Assert.AreEqual(isEditor || isIL2CPP ? sceneManager.GetActiveScene().Name : null, unityProtocol!.ActiveSceneName);
     }
 }